Make a common NullableConverter<T> for System.Text.Json
Motivation
----------
Various extensions, including Yardarm.NodaTime, need to register simple
custom converters for nullable value types that wrap an inner converter
and handle nulls. To prevent code duplication, make this a common
internal member available with the Yardarm.SystemTextJson extension.

Modifications
-------------
- Move NodaNullableConverter to NullableConverter in the STJ client
- Add references and InternalsVisibleTo to support testing
