+++ src/main/Yardarm.SystemTextJson.Client/Properties/ClientAssemblyInfo.cs
@@ -2,4 +2,5 @@
 
 #if FORTESTS
 [assembly: InternalsVisibleTo("Yardarm.SystemTextJson.UnitTests")]
+[assembly: InternalsVisibleTo("Yardarm.NodaTime.Client")]
 #endif
